+++ app/lib/shared/versions.dart
@@ -16,7 +16,7 @@ final RegExp runtimeVersionPattern = RegExp(r'\d{4}\.\d{2}\.\d{2}');
 /// Increment the version when a change is significant enough to trigger
 /// reprocessing, including: version change in pana, dartdoc, or the SDKs,
 /// or when an feature or bugfix should be picked up by the analysis ASAP.
-const String runtimeVersion = '2019.07.15';
+const String runtimeVersion = '2019.08.13';
 final Version semanticRuntimeVersion = Version.parse(runtimeVersion);
 
 /// The version which marks the earliest version of the data which we'd like to
@@ -25,6 +25,7 @@ final Version semanticRuntimeVersion = Version.parse(runtimeVersion);
 ///
 /// Make sure that at least two versions are kept here as the next candidates
 /// when the version switch happens:
+/// - 2019.08.13
 /// - 2019.07.10
 /// - 2019.07.01
 /// - 2019.06.24
@@ -39,7 +40,7 @@ final String toolEnvSdkVersion = '2.4.0';
 final String panaVersion = '0.12.19';
 final Version semanticPanaVersion = Version.parse(panaVersion);
 
-final String flutterVersion = '1.7.8+hotfix.3';
+final String flutterVersion = '1.7.8+hotfix.4';
 final Version semanticFlutterVersion = Version.parse(flutterVersion);
 
 // keep in-sync with pkg/pub_dartdoc/pubspec.yaml