+++ CHANGELOG.md
@@ -4,7 +4,8 @@ AppEngine version, listed here to ease deployment and troubleshooting.
 ## Next Release (replace with git tag when deployed)
 
  * Upgraded dartdoc to `0.28.4` (upgraded analyzer).
- * Bumped runtimeVersion to `2019.07.15`.
+ * Upgraded Flutter SDK to `1.7.8+hotfix.4`.
+ * Bumped runtimeVersion to `2019.08.13`.
  * Increased HSTS duration to a year.
 
 ## `20190711t114908-all`